requirements.txt
dnspython==2.1.0
influxdb==5.3.1
pythonping==1.1.0